feat: promote estimator custom reducers from experimental
