+++ harness/determined/estimator/__init__.py
@@ -1,15 +1,16 @@
 from determined.estimator._callback import RunHook
+from determined.estimator._reducer import (
+    MetricReducer,
+    _SimpleMetricReducer,
+    _DistributedMetricMaker,
+    _EstimatorReducerContext,
+)
 from determined.estimator._estimator_context import (
     EstimatorNativeContext,
     EstimatorContext,
     EstimatorExperimentalContext,
     EstimatorTrialContext,
     ServingInputReceiverFn,
-)
-from determined.estimator._reducer import (
-    MetricReducer,
-    _SimpleMetricReducer,
-    _DistributedMetricMaker,
 )
 from determined.estimator._util import (
     _cleanup_after_train_step,
