+++ harness/determined/estimator/_reducer.py
@@ -1,5 +1,5 @@
 import abc
-from typing import Any, Callable, List, Sequence, Tuple
+from typing import Any, Callable, List, Sequence, Tuple, Union
 
 import numpy as np
 import tensorflow as tf
@@ -43,7 +43,7 @@ class MetricReducer:
             ...
             if mode == tf.estimator.ModeKeys.EVAL:
 
-                my_avg_prediction = context.experimental.make_metric(
+                my_avg_prediction = context.make_metric(
                      metric=predictions, reducer=MyAvgMetricReducer(), numpy_dtype=np.float32
                 )
 
@@ -53,7 +53,7 @@ class MetricReducer:
                     eval_metric_ops={"my_avg_prediction": my_avg_prediction},
                 )
 
-    See also: :func:`determined.estimator.EstimatorExperimentalContext.make_metric`.
+    See also: :func:`context.make_metric() <determined.estimator.EstimatorContext.make_metric>`.
     """
 
     @abc.abstractmethod
@@ -115,6 +115,120 @@ class _SimpleMetricReducer(MetricReducer):
     def cross_slot_reduce(self, per_slot_metrics: List[List[Any]]) -> Any:
         flat_metrics = [item for sublist in per_slot_metrics for item in sublist]
         return self.reduce_fn(flat_metrics)
+
+
+def default_allgather_fn(metrics: Any) -> List:
+    """
+    A noop allgather implementation to ensure that custom reducers work outside of Determined.
+    """
+    return [metrics]
+
+
+class _EstimatorReducerContext:
+    """
+    Context class that contains only the reducer-related information.
+    """
+
+    def __init__(self) -> None:
+        self._allgather_fn = default_allgather_fn  # type: Callable[[Any], List]
+        # allgather is not parallelizable, so we have to strictly order how they are placed in the
+        # graph via tf.control_dependencies().
+        self._allgather_ops = []  # type: List[tf.Operation]
+
+    def _set_allgather_fn(self, fn: Callable[[Any], List]) -> None:
+        self._allgather_fn = fn
+
+    def _build_allgather_op(self, build_op_fn: Callable[[], tf.Operation]) -> tf.Operation:
+        """Build an op that uses allgather in a way that is safely sequentialized."""
+
+        with tf.compat.v1.control_dependencies(self._allgather_ops):
+            new_op = build_op_fn()
+        self._allgather_ops.append(new_op)
+        return new_op
+
+    def _reset_allgather_ops(self) -> None:
+        """Every Estimator evaluation happens on a clean graph, so forget the old operations."""
+        self._allgather_ops = []
+
+    def make_metric(
+        self,
+        metric: Any,
+        reducer: Union[Callable[[List[Any]], Any], "estimator.MetricReducer"],
+        numpy_dtype: Any,
+    ) -> Tuple[tf.Operation, tf.Operation]:
+        """
+        Return an estimator-compatible validation metric which will be calculated properly, even
+        during distributed evaluation.
+
+        During distributed evaluation, many types of metrics calculated via ``tf.metrics`` or
+        ``tf.keras.metrics`` cannot be aggregated properly from the per-slot final metrics
+        calculated by each separate Estimator replica. One example is ``tf.metrics.auc``, where
+        the ROC AUC calculated over predictions and labels from a full dataset cannot be derived
+        from the individual ROC AUC metrics evaluated over several shards of a dataset.
+
+        Determined solves this problem by offering customizable metrics which are
+        Estimator-compatible.  For example, ROC AUC could be properly calculated during distributed
+        evaluation by calling ``sklearn.metrics.roc_auc_score`` in a custom ``reducer`` function
+        passed to ``make_metric``.
+
+        The ``metric`` input can be a tensor, a list of tensors, or a dictionary of tensors.
+        Nested structures are not supported.
+
+        The ``reducer`` should be either a single function that can calculate the metric from a
+        list of the per-batch values of ``metric``, or it can be an instance of a
+        :class:`det.estimator.MetricReducer<determined.estimator.MetricReducer>`.
+
+        The ``numpy_dtype`` must be a numpy dtype.  It is used internally to determined the output
+        type of the TensorFlow ``py_func`` to report the final metric result to the Estimator API.
+        The format of ``numpy_dtype`` should be anything that ``np.dtype()`` accepts.
+
+        The primary motivation for passing a function as the reducer is simplicity. Metrics from
+        all batches will be buffered in memory and passed over the network where they will be
+        reduced all at once. This introduces some overhead, but it is likely unnoticeable for
+        scalar metrics or on validation datasets of small or medium size. This single function
+        strategy may also be desirable for quick prototyping or for calculating metrics that are
+        difficult or impossible to calculate incrementally.
+
+        The primary motivation for passing a ``det.estimator.MetricsReducer`` as the reducer is
+        performance. ``det.estimator.MetricsReducer`` allows the user to incrementally calculate
+        the partial metric on each slot, taking advantage of distributed computation, minimizing
+        memory usage, and minimizing the network communication before the final
+        ``cross_slot_reduce`` operation.
+
+        Evaluation performance may be improved by precomputing as much as possible in the graph so
+        that less computation on the ``metric`` value is required within the reducer.
+
+        Example usage where ``reducer`` is a function:
+
+        .. code-block:: python
+
+           def my_mean_reducer(all_batch_metrics):
+               # Use hstack in case not all batches are equal length.
+               return np.mean(np.hstack(all_batch_metrics))
+
+           def my_estimator_model_function(features, labels, mode):
+               ...
+               if mode == tf.estimator.ModeKeys.EVAL:
+
+                   my_avg_prediction = context.make_metric(
+                        metric=predictions, reducer=my_mean_reducer, numpy_dtype=np.float32
+                   )
+
+                   return tf.estimator.EstimatorSpec(
+                       mode,
+                       loss=loss,
+                       eval_metric_ops={"my_avg_prediction": my_avg_prediction},
+                   )
+        """
+        if isinstance(reducer, estimator.MetricReducer):
+            return estimator._DistributedMetricMaker(
+                self, metric, reducer, numpy_dtype
+            ).make_metric()
+
+        simple_reducer = estimator._SimpleMetricReducer(reducer)
+        return estimator._DistributedMetricMaker(
+            self, metric, simple_reducer, numpy_dtype
+        ).make_metric()
 
 
 def _deconstruct_metric(metric: Any) -> Tuple[Sequence, Any]:
@@ -184,7 +298,7 @@ class _DistributedMetricMaker:
 
     def __init__(
         self,
-        context: estimator.EstimatorExperimentalContext,
+        context: _EstimatorReducerContext,
         metric: Any,
         reducer: MetricReducer,
         numpy_dtype: Any,
@@ -209,7 +323,7 @@ class _DistributedMetricMaker:
         return tf.compat.v1.py_func(self._update, self.update_args, [])
 
     def _value(self) -> Any:
-        allgathered = self.context.allgather_metrics(self.last_accumulate)
+        allgathered = self.context._allgather_fn(self.last_accumulate)
         value = self.reducer.cross_slot_reduce(allgathered)
         return np.array(value).astype(self.np_dtype)
